requirements.txt
Flask
python-dotenv
Flask-Session
openai
requests
